mode done

+++ app/src/main/java/com/technosales/net/buslocationannouncement/APIToken/CustomAuthenticator.java
@@ -40,7 +40,7 @@ import static com.technosales.net.buslocationannouncement.utils.UtilStrings.TICK
 
 
 public class CustomAuthenticator implements Authenticator {
-    int a;
+    int bearer=0;
       private TokenManager tokenManager;
         private static CustomAuthenticator INSTANCE;
         private CustomAuthenticator(TokenManager tokenManager){
@@ -57,44 +57,51 @@ public class CustomAuthenticator implements Authenticator {
         @Nullable
         @Override
         public Request authenticate(Route route, Response response) throws IOException {
-           HelperModel.Token token = tokenManager.getToken();
-           a++;
-            Log.i("TAG", "authenticate: "+route.toString());
-            if (response.code() == 401) {
-                HttpLoggingInterceptor interceptor = new HttpLoggingInterceptor();
-                interceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
+            HelperModel.Token token = tokenManager.getToken();
+            bearer++;
+            Log.i("TAG", "authenticate: " + bearer);
+            if (bearer == 1) {
+                if (response.code() == 401) {
+                    HttpLoggingInterceptor interceptor = new HttpLoggingInterceptor();
+                    interceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
 
-                OkHttpClient client = new OkHttpClient.Builder()
-                        .connectTimeout(100, TimeUnit.SECONDS)
-                        .readTimeout(100, TimeUnit.SECONDS)
-                        .addInterceptor(interceptor)
-                        .build();
-                Retrofit.Builder builder = new Retrofit.Builder()
-                        .client(client)
-                        .baseUrl(TICKET_URL_NEW)
-                        .addConverterFactory(GsonConverterFactory.create());
+                    OkHttpClient client = new OkHttpClient.Builder()
+                            .connectTimeout(100, TimeUnit.SECONDS)
+                            .readTimeout(100, TimeUnit.SECONDS)
+                            .addInterceptor(interceptor)
+                            .build();
+                    Retrofit.Builder builder = new Retrofit.Builder()
+                            .client(client)
+                            .baseUrl(TICKET_URL_NEW)
+                            .addConverterFactory(GsonConverterFactory.create());
 
-                Log.i("TAG", "authenticate: "+ token.getRefreshToken());
-                Retrofit retrofit = builder.build();
-               RetrofitInterface apiInterface = retrofit.create(RetrofitInterface.class);
-                Call<HelperModel.Token> call = apiInterface.refresh(token.getRefreshToken());
-                retrofit2.Response< HelperModel.Token> res = call.execute();
-                if (res != null && res.code() == 200) {
-                    if (res.isSuccessful()) {
-                        HelperModel.Token newToken = res.body();
-                        tokenManager.saveToken(newToken);
-                        Log.i("TAG", "authenticate: " + res.body().getRefreshToken());
-                        return response.request().newBuilder().header("Authorization", "Bearer " + res.body().getAccessToken()).build();
+                    Log.i("TAG", "authenticate: " + token.getRefreshToken());
+                    Retrofit retrofit = builder.build();
+                    RetrofitInterface apiInterface = retrofit.create(RetrofitInterface.class);
+                    Call<HelperModel.Token> call = apiInterface.refresh(token.getRefreshToken());
+                    retrofit2.Response<HelperModel.Token> res = call.execute();
+                    if (res != null && res.code() == 200) {
+                        if (res.isSuccessful()) {
+                            bearer=0;
+                            HelperModel.Token newToken = res.body();
+                            tokenManager.saveToken(newToken);
+                            Log.i("TAG", "authenticate: " + res.body().getRefreshToken());
+                            return response.request().newBuilder().header("Authorization", "Bearer " + res.body().getAccessToken()).build();
+                        }
+                    } else if (res.code() == 401) {
+                        bearer=0;
+                        SharedPreferences preferences = TicketBusApp.getContext().getSharedPreferences(SHARED_PREFERENCES_HELPER, Context.MODE_PRIVATE);
+                        SharedPreferences.Editor editor = preferences.edit();
+                        editor.clear();
+                        editor.apply();
+                        TicketBusApp.getContext().startActivity(new Intent(TicketBusApp.getContext(), HelperLogin.class));
                     }
-                } else if(res.code() == 401){
-                    SharedPreferences preferences = TicketBusApp.getContext().getSharedPreferences(SHARED_PREFERENCES_HELPER, Context.MODE_PRIVATE);
-                    SharedPreferences.Editor editor = preferences.edit();
-                    editor.clear();
-                    editor.apply();
-                    TicketBusApp.getContext().startActivity(new Intent(TicketBusApp.getContext(),HelperLogin.class));
                 }
+                return null;
+            }else {
+                return null;
             }
-            return  null;
+
         }
 
     }
